requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements/requirements.txt subset.in
#
backports-zoneinfo==0.2.1
    # via
    #   -c requirements/requirements_ci.txt
    #   pytz-deprecation-shim
    #   tzlocal
certifi==2021.10.8
    # via
    #   -c requirements/requirements_ci.txt
    #   requests
cfg-load==0.9.0
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
charset-normalizer==2.0.12
    # via
    #   -c requirements/requirements_ci.txt
    #   requests
deprecated==1.2.13
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
dill==0.3.4
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
idna==3.3
    # via
    #   -c requirements/requirements_ci.txt
    #   requests
mongomock==4.0.0
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
mpu[io]==0.23.1
    # via
    #   -c requirements/requirements_ci.txt
    #   cfg-load
numpy==1.21.5
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
    #   pandas
    #   scipy
packaging==21.3
    # via
    #   -c requirements/requirements_ci.txt
    #   mongomock
pandas==1.3.5
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
pymongo==4.0.1
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
pyparsing==3.0.7
    # via
    #   -c requirements/requirements_ci.txt
    #   packaging
python-dateutil==2.8.2
    # via
    #   -c requirements/requirements_ci.txt
    #   pandas
pytz==2021.3
    # via
    #   -c requirements/requirements_ci.txt
    #   cfg-load
    #   mpu
    #   pandas
pytz-deprecation-shim==0.1.0.post0
    # via
    #   -c requirements/requirements_ci.txt
    #   tzlocal
pyyaml==6.0
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
    #   cfg-load
requests==2.27.1
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
    #   cfg-load
scipy==1.5.4
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
sentinels==1.0.0
    # via
    #   -c requirements/requirements_ci.txt
    #   mongomock
six==1.16.0
    # via
    #   -c requirements/requirements_ci.txt
    #   cfg-load
    #   mongomock
    #   python-dateutil
typeguard==2.13.3
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
typing-extensions==4.1.1
    # via
    #   -c requirements/requirements_ci.txt
    #   -r subset.in
tzdata==2021.5
    # via
    #   -c requirements/requirements_ci.txt
    #   pytz-deprecation-shim
tzlocal==4.1
    # via
    #   -c requirements/requirements_ci.txt
    #   mpu
urllib3==1.26.8
    # via
    #   -c requirements/requirements_ci.txt
    #   requests
wrapt==1.13.3
    # via
    #   -c requirements/requirements_ci.txt
    #   deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
